Solves for a tabular policy based on histories with MLE.
The Maximum Likelihood Estimator of the policy for each state is
pi(state, act) = c(state, act) / [sum(c(state, act)) for all act in actions],
Where c(state, act) is the numer of times `act` was observed in the input @ref
histories.

If this process results in a zero likelihood for an action, it is replaced
by sys.float_info.min.