Implement gradientAscentGaussianTheta
During gradient ascent, the elements of the parameter vector theta are
assumed to be sampled i.i.d. from a gaussian distribution with mean
theta_mu[i] and standard deviation theta_sigma[i]. Gradient ascent is
performed on the means and std_deviations.

The infered policy can be reconstructed with the Boltzman policy using
the returned theta_mu vector and the uncertainty about the mean is
returned as the theta_sigma vector.