Simplify value iteration termination and printing
Value iteration now terminates at the specified epsilon instead of at a
threshold that is a function of epsilon and gamma. This makes it easier
to configure how long value iteration runs.

Value iteration accepts a print-iterations flag to allow the user to
see the iteration improvment/timing without printing the state-values
or policy, which often clog up the console.