Prevent integer multiplication overflow errors by correctly setting array data-types
Previously, some arrays were configured to be numpy.int8, and others
were initialized with a data-type based off of the highest number they
were supposed to hold. However, when these arrays were operated on with
integer arrays, there was frequent overflows that would require the
next largest data-type. Since this level of optimization is largely
unnecessary this commit generalizes those data-types, or lets them be
automatically selected as the default numpy.int64.