Add option for single-agent MDP to be built without DRA product
This option was previously hard-coded into experiment_configs.py. Now
the flag is passed into makeGridMDPxDRA(). This affects the format of
the states in the MDP. In an MDPxDRA with DRA states, the states are
formatted as tuples (grid_cell(s), dra_state), where grid_cell(s)
is a tuple for each agent's state grid_cells = (s_1, s_2) for two agents
and = (s_1) for a single agent. In the case that a ProductMDPxDRA()
object is built without DRA states, en example state is formatted as
((3,),) representing thta a singlular agent is at grid cell 3.